+++ src/editor/pages/parts/viewport3D.cpp
@@ -21,6 +21,7 @@
 #include "../../undoRedo.h"
 #include "../../selectionUtils.h"
 
+#include "../../../utils/logger.h"
 namespace
 {
   constinit uint32_t nextPassId{0};
@@ -349,7 +350,7 @@ void Editor::Viewport3D::draw()
   mousePos.x -= screenPos.x;
   mousePos.y -= vpOffsetY;
 
-  float moveSpeed = 120.0f * deltaTime;
+  float moveSpeed = ctx.moveSpeed * deltaTime;
 
   bool mouseHeldLeft = ImGui::IsMouseDown(ImGuiMouseButton_Left);
   bool mouseHeldRight = ImGui::IsMouseDown(ImGuiMouseButton_Right);
@@ -478,19 +479,20 @@ void Editor::Viewport3D::draw()
     //instead, we have to rely on the fact that trackpads move in fractional amounts
     glm::vec2 wheel = glm::vec2(io.MouseWheelH, io.MouseWheel);
     bool usesWheel = wheel != glm::vec2{0,0};
-
+    
     if(usesWheel)
     {
-      if (std::abs(wheel.x) == 1 || std::abs(wheel.y) == 1) {
+      if (std::fmod(std::abs(wheel.x), 1.0f) == 0 && std::fmod(std::abs(wheel.y), 1.0f) == 0) {
         //actual wheel or pinch gesture
-        float wheelSpeed = (isShiftDown ? 4.0f : 1.0f) * 30.0f;
+        float wheelSpeed = (isShiftDown ? 4.0f : 1.0f) * ctx.zoomSpeed;
         camera.zoomSpeed += wheel.y * wheelSpeed;
       } else {
+        if (ctx.invertWheelY) wheel.y *= -1;
         //two finger swipe on trackpad
         if (isShiftDown) {
-          camera.moveDelta(wheel * -30.0f);
+          camera.moveDelta(wheel * ctx.panSpeed);
         } else {
-          camera.orbitDelta(wheel * 10.0f);
+          camera.orbitDelta(wheel * ctx.lookSpeed);
         }
       }
     }
@@ -506,6 +508,7 @@ void Editor::Viewport3D::draw()
   //ImGui::Text("Viewport: %f | %f | %08X", mousePos.x, mousePos.y, ctx.selObjectUUID);
 
   constexpr const char* const GIZMO_LABELS[3] = {ICON_MDI_CURSOR_MOVE, ICON_MDI_ROTATE_360, ICON_MDI_ARROW_EXPAND};
+  constexpr const char* const GIZMO_TOOLTIPS[3] = {"Translate", "Rotate", "Scale"};
   for (int i=0; i<3; ++i) {
     if (ConnectedToggleButton(
       GIZMO_LABELS[i],
@@ -515,6 +518,7 @@ void Editor::Viewport3D::draw()
     )) {
       gizmoOp = i;
     }
+    ImGui::SetItemTooltip("%s", GIZMO_TOOLTIPS[i]);
   }
 
   ImGui::SameLine();
@@ -522,6 +526,7 @@ void Editor::Viewport3D::draw()
   if (ConnectedToggleButton(ICON_MDI_WEB, isTransWorld, true, true, ImVec2(32,24))) {
     isTransWorld = !isTransWorld;
   }
+  ImGui::SetItemTooltip("Show %s Space", isTransWorld ? "Local" : "World");
 
   ImGui::SameLine();
   ImGui::SetCursorPosX(ImGui::GetCursorPosX() + 12);
@@ -529,18 +534,21 @@ void Editor::Viewport3D::draw()
   if(ConnectedToggleButton(ICON_MDI_GRID, showGrid, true, true, ImVec2(32,24))) {
     showGrid = !showGrid;
   }
+  ImGui::SetItemTooltip("%s Grid", showGrid ? "Hide" : "Show");
 
   ImGui::SameLine();
   ImGui::SetCursorPosX(ImGui::GetCursorPosX() - 4);
   if(ConnectedToggleButton(ICON_MDI_LANDSLIDE_OUTLINE, showCollMesh, true, true, ImVec2(32,24))) {
     showCollMesh = !showCollMesh;
   }
+  ImGui::SetItemTooltip("%s Collision Mesh", showCollMesh ? "Hide" : "Show");
 
   ImGui::SameLine();
   ImGui::SetCursorPosX(ImGui::GetCursorPosX() - 4);
   if(ConnectedToggleButton(ICON_MDI_CYLINDER, showCollObj, true, true, ImVec2(32,24))) {
     showCollObj = !showCollObj;
   }
+  ImGui::SetItemTooltip("%s Collision Bodies", showCollObj ? "Hide" : "Show");
 
   ImGui::SetCursorPosY(currPos.y + BAR_HEIGHT);