+++ src/main.cpp
@@ -37,6 +37,7 @@
 #include <cctype>
 
 #include "editor/undoRedo.h"
+#include "editor/actions.h"
 
 Context ctx{};
 constinit SDL_GPUSampler *texSamplerRepeat{nullptr};
@@ -274,6 +275,7 @@ int main(int argc, char** argv)
 
     // Main loop
     bool done = false;
+    float lastPinch;
     while(!done) {
 
       auto frameStart = SDL_GetTicksNS();
@@ -281,6 +283,19 @@ int main(int argc, char** argv)
       SDL_Event event;
       while (SDL_PollEvent(&event))
       {
+        //convert pinch events to whole number mouse wheel events to mimic windows 
+        if (event.type == SDL_EVENT_PINCH_BEGIN) {
+          lastPinch = 1;
+        } else if (event.type == SDL_EVENT_PINCH_UPDATE) {
+
+          float y = event.pinch.scale - lastPinch;
+          if (y > 0) y = 1;
+          else if (y < 0) y = -1;
+          lastPinch = event.pinch.scale;
+          io.AddMouseSourceEvent(ImGuiMouseSource_Mouse);
+          io.AddMouseWheelEvent(0, y);
+        }
+
         ImGui_ImplSDL3_ProcessEvent(&event);
 
         bool closeRequested = event.type == SDL_EVENT_QUIT;
@@ -297,9 +312,11 @@ int main(int argc, char** argv)
         }
 
         if(event.type == SDL_EVENT_WINDOW_MOVED || event.type == SDL_EVENT_WINDOW_RESIZED
-          || event.type == SDL_EVENT_WINDOW_RESTORED || event.type == SDL_EVENT_WINDOW_SHOWN)
-        {
+          || event.type == SDL_EVENT_WINDOW_RESTORED || event.type == SDL_EVENT_WINDOW_SHOWN) {
           editorWindow.trackGeometry();
+        } else if (event.type == SDL_EVENT_DROP_FILE) {
+          std::string droppedPath = event.drop.data;
+          if (droppedPath.ends_with(".p64proj"))Editor::Actions::call(Editor::Actions::Type::PROJECT_OPEN, droppedPath);
         }
 
         // @TODO: refactor into generic actions with keybinds